Replaced most currency values with BigDecimal for precision. Prepared for adding new view on home screen

+++ app/src/main/java/com/github/kwasow/archipelago/data/Transaction.kt
@@ -1,11 +1,12 @@
 package com.github.kwasow.archipelago.data
 
 import java.io.Serializable
+import java.math.BigDecimal
 import java.util.Date
 
 data class Transaction(
-    var date: Date,
-    var name: String,
-    var amount: Double,
-    var details: String
+        var date: Date,
+        var name: String,
+        var amount: BigDecimal,
+        var details: String
 ) : Serializable
